new machine who dis
* adds lots of missing config from new machine startup
* creates npm installation script
* extends coc stuff with extensions and config file
* better aliasing in fish for external calls

+++ config/fish/config.fish
@@ -5,17 +5,32 @@ else
   alias la='ls -la'
 end
 alias l="ls -laFGgohq"
-alias fsh='vim ~/.config/fish/'
-alias vrc='vim ~/.vim/settings/'
-alias tmc='vim ~/.tmux.conf'
-alias gco='git checkout'
-alias gs='git status'
-alias ga='git add'
-alias t='todolist'
-alias nats="docker run -d -p 4222:4222 -p 6222:6222 -p 8222:8222 nats:latest"
-alias vim=nvim
-alias fuck_pg='pg_ctl -D /usr/local/var/postgres -l /usr/local/var/postgres/server.log start'
-alias fm='fzf | xargs rm -rfi'
+if type -q vim
+  alias fsh='vim ~/.config/fish/'
+  alias vrc='vim ~/.vim/settings/'
+  alias tmc='vim ~/.tmux.conf'
+end
+if type -q git
+  alias gco='git checkout'
+  alias gs='git status'
+  alias ga='git add'
+end
+if type -q todolist
+  alias t='todolist'
+end
+if type -q docker
+  alias nats="docker run -d -p 4222:4222 -p 6222:6222 -p 8222:8222 nats:latest"
+end
+if type -q docker
+  alias vim=nvim
+  alias vimdiff="nvim -d"
+end
+if type -q pg_ctl
+  alias fuck_pg='pg_ctl -D /usr/local/var/postgres -l /usr/local/var/postgres/server.log start'
+end
+if type -q fzf
+  alias fm='fzf | xargs rm -rfi'
+end
 status --is-interactive; and source (rbenv init -|psub)
 status --is-interactive; and source (pyenv init -|psub)
 status --is-interactive; and source (nodenv init -|psub)
